requirements.txt
pystac-client
planetary-computer
stackstac
rioxarray
xarray
pandas
pyyaml
matplotlib
dask
rasterio
requests
folium
streamlit-folium
geopy